requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
pydantic>=2.0
pvlib>=0.11
pyomo>=6.7
highspy>=1.7
matplotlib>=3.7
jinja2>=3.1
pytest>=7.4
